+++ src/lyskal/module8/wokers/entities/Person.java
@@ -1,4 +1,6 @@
-package lyskal.module8.wokers;
+package lyskal.module8.wokers.entities;
+
+import lyskal.module8.wokers.servises.Validator;
 
 public abstract class Person {
 	private String _name;
